+++ packages/widgetbook/lib/src/addons/theme_addon/theme_addon.dart
@@ -2,6 +2,7 @@ import 'package:flutter/widgets.dart';
 
 import '../../core/core.dart';
 import '../../fields/fields.dart';
+import '../../routing/routing.dart';
 
 typedef ThemeBuilder<T> =
     Widget Function(
@@ -48,12 +49,12 @@ class ThemeAddon<T> extends Addon<T> {
   }
 
   @override
-  T valueFromQueryGroup(Map<String, String> group) {
+  T valueFromQueryGroup(QueryGroup group) {
     return valueOf<T>('name', group)!;
   }
 
   @override
-  Map<String, String> valueToQueryGroup(T value) {
+  QueryGroup valueToQueryGroup(T value) {
     return {'name': paramOf('name', value)};
   }