+++ packages/widgetbook/lib/src/addons/viewport_addon/viewport_addon.dart
@@ -2,6 +2,7 @@ import 'package:flutter/widgets.dart' hide Viewport;
 
 import '../../core/core.dart';
 import '../../fields/fields.dart';
+import '../../routing/routing.dart';
 import 'viewport.dart';
 import 'viewport_data.dart';
 import 'viewports/viewports.dart';
@@ -40,12 +41,12 @@ class ViewportAddon extends Addon<ViewportData> {
   }
 
   @override
-  ViewportData valueFromQueryGroup(Map<String, String> group) {
+  ViewportData valueFromQueryGroup(QueryGroup group) {
     return valueOf<ViewportData>('name', group)!;
   }
 
   @override
-  Map<String, String> valueToQueryGroup(ViewportData value) {
+  QueryGroup valueToQueryGroup(ViewportData value) {
     return {'name': paramOf('name', value)};
   }